+++ examples/mystery-sentence/src/mutation.cpp
@@ -6,7 +6,7 @@
 Mutation::Mutation(Generator* generator) : generator(generator) {}
 
 void Mutation::operator()(std::string& solution) const {
-    for (int i = 0; i < solution.length(); ++i) {
+    for (size_t i = 0; i < solution.length(); ++i) {
         if (GenericGenetics::Random::getInt(0, 99) < 10) {
             solution[i] =
                 this->generator->alphanum[GenericGenetics::Random::getInt(
